[ISSUE-109] Attachement should be VertxGen object
