+++ src/main/java/io/vertx/ext/mail/MailAttachment.java
@@ -16,56 +16,30 @@
 
 package io.vertx.ext.mail;
 
-import io.vertx.codegen.annotations.DataObject;
-import io.vertx.codegen.annotations.GenIgnore;
+import io.vertx.codegen.annotations.Fluent;
+import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.MultiMap;
 import io.vertx.core.buffer.Buffer;
-import io.vertx.core.http.CaseInsensitiveHeaders;
 import io.vertx.core.json.JsonObject;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
+import io.vertx.ext.mail.impl.MailAttachmentImpl;
 
 /**
  * Represent a mail attachment that can be used in a MailMessage.
  *
  * @author <a href="http://oss.lehmann.cx/">Alexander Lehmann</a>
+ * @author <a href="mailto:aoingl@gmail.com">Lin Gao</a>
  */
 
-@DataObject
-public class MailAttachment {
-
-  private Buffer data;
-  private String name;
-  private String contentType;
-  private String disposition;
-  private String description;
-  private String contentId;
-  private MultiMap headers;
+@VertxGen
+public interface MailAttachment {
 
   /**
    * construct an empty MailAttachment object that can be filled with the
    * setters
    */
-  public MailAttachment() {
-  }
-
-  /**
-   * create a copy of a MailAttachment object
-   *
-   * @param other object to be copied
-   */
-  public MailAttachment(final MailAttachment other) {
-    Objects.requireNonNull(other);
-    this.data = other.data == null ? null : other.data.copy();
-    this.name = other.name;
-    this.contentType = other.contentType;
-    this.disposition = other.disposition;
-    this.description = other.description;
-    this.description = other.description;
-    this.contentId = other.contentId;
-    this.headers = other.headers == null ? null : new CaseInsensitiveHeaders().addAll(other.headers);
+  static MailAttachment create() {
+    return new MailAttachmentImpl();
   }
 
   /**
@@ -73,28 +47,23 @@ public class MailAttachment {
    *
    * @param json object to be copied
    */
-  public MailAttachment(final JsonObject json) {
-    Objects.requireNonNull(json);
-    this.data = json.getBinary("data") == null ? null : Buffer.buffer(json.getBinary("data"));
-    this.name = json.getString("name");
-    this.contentType = json.getString("contentType");
-    this.disposition = json.getString("disposition");
-    this.description = json.getString("description");
-    this.contentId = json.getString("contentId");
-    JsonObject headers = json.getJsonObject("headers");
-    if (headers != null) {
-      this.headers = Utils.jsonToMultiMap(headers);
-    }
-  }
+  static MailAttachment create(JsonObject json) {return new MailAttachmentImpl(json); }
+
+   /**
+    * create a copy of a MailAttachment object
+    *
+    * @param other object to be copied
+    */
+  static MailAttachment create(MailAttachment other) {
+    return new MailAttachmentImpl(other);
+   }
 
   /**
    * get the data
    *
    * @return the data
    */
-  public Buffer getData() {
-    return data;
-  }
+  Buffer getData();
 
   /**
    * set the data
@@ -102,19 +71,15 @@ public class MailAttachment {
    * @param data Buffer of bytes to be used at attachment
    * @return this to be able to use it fluently
    */
-  public MailAttachment setData(final Buffer data) {
-    this.data = data;
-    return this;
-  }
+  @Fluent
+  MailAttachment setData(Buffer data);
 
   /**
    * get the name
    *
    * @return the name
    */
-  public String getName() {
-    return name;
-  }
+  String getName();
 
   /**
    * set the name
@@ -125,19 +90,15 @@ public class MailAttachment {
    * i.e. usually a local filename without path
    * this can be set to "" to omit the filename attribute
    */
-  public MailAttachment setName(final String name) {
-    this.name = name;
-    return this;
-  }
+  @Fluent
+  MailAttachment setName(String name);
 
   /**
    * get the Content-Type
    *
    * @return the contentType
    */
-  public String getContentType() {
-    return contentType;
-  }
+  String getContentType();
 
   /**
    * set the Content-Type
@@ -145,19 +106,15 @@ public class MailAttachment {
    * @param contentType the contentType
    * @return this to be able to use it fluently
    */
-  public MailAttachment setContentType(final String contentType) {
-    this.contentType = contentType;
-    return this;
-  }
+  @Fluent
+  MailAttachment setContentType(String contentType);
 
   /**
    * get the disposition field
    *
    * @return the disposition
    */
-  public String getDisposition() {
-    return disposition;
-  }
+  String getDisposition();
 
   /**
    * set the disposition field to be used in the attachment
@@ -165,19 +122,15 @@ public class MailAttachment {
    * @param disposition the disposition
    * @return this to be able to use it fluently
    */
-  public MailAttachment setDisposition(final String disposition) {
-    this.disposition = disposition;
-    return this;
-  }
+  @Fluent
+  MailAttachment setDisposition(String disposition);
 
   /**
    * get the description field
    *
    * @return the description
    */
-  public String getDescription() {
-    return description;
-  }
+  String getDescription();
 
   /**
    * set the description field to be used in the attachment
@@ -185,19 +138,15 @@ public class MailAttachment {
    * @param description the description
    * @return this to be able to use it fluently
    */
-  public MailAttachment setDescription(final String description) {
-    this.description = description;
-    return this;
-  }
+  @Fluent
+  MailAttachment setDescription(String description);
 
   /**
    * get the Content-ID field
    *
    * @return the content id
    */
-  public String getContentId() {
-    return contentId;
-  }
+  String getContentId();
 
   /**
    * set the Content-ID field to be used in the attachment
@@ -205,10 +154,8 @@ public class MailAttachment {
    * @param contentId the content id
    * @return this to be able to use it fluently
    */
-  public MailAttachment setContentId(final String contentId) {
-    this.contentId = contentId;
-    return this;
-  }
+  @Fluent
+  MailAttachment setContentId(String contentId);
 
   /**
    * Add an header to this attachment.
@@ -217,25 +164,15 @@ public class MailAttachment {
    * @param value  the header value
    * @return  a reference to this, so the API can be used fluently
    */
-  public MailAttachment addHeader(String key, String value) {
-    if (headers == null) {
-      headers = new CaseInsensitiveHeaders();
-    }
-    Objects.requireNonNull(key, "no null key accepted");
-    Objects.requireNonNull(value, "no null value accepted");
-    headers.add(key, value);
-    return this;
-  }
+  @Fluent
+  MailAttachment addHeader(String key, String value);
 
   /**
    * Get the headers to be added for this attachment.
    *
    * @return the headers
    */
-  @GenIgnore
-  public MultiMap getHeaders() {
-    return headers;
-  }
+  MultiMap getHeaders();
 
   /**
    * Set the headers to be added for this attachment.
@@ -243,52 +180,14 @@ public class MailAttachment {
    * @param headers the headers to be added
    * @return this to be able to use it fluently
    */
-  @GenIgnore
-  public MailAttachment setHeaders(final MultiMap headers) {
-    this.headers = headers;
-    return this;
-  }
+  @Fluent
+  MailAttachment setHeaders(MultiMap headers);
 
   /**
    * convert this object to JSON representation
    *
    * @return the JSON object
    */
-  public JsonObject toJson() {
-    JsonObject json = new JsonObject();
-    if (data != null) {
-      json.put("data", data.getBytes());
-    }
-    Utils.putIfNotNull(json, "name", name);
-    Utils.putIfNotNull(json, "contentType", contentType);
-    Utils.putIfNotNull(json, "disposition", disposition);
-    Utils.putIfNotNull(json, "description", description);
-    Utils.putIfNotNull(json, "contentId", contentId);
-    if (headers != null) {
-      json.put("headers", Utils.multiMapToJson(headers));
-    }
-    return json;
-  }
-
-  private List<Object> getList() {
-    return Arrays.asList(data, name, disposition, description, contentId, headers);
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (!(o instanceof MailAttachment)) {
-      return false;
-    }
-    final MailAttachment attachment = (MailAttachment) o;
-    return getList().equals(attachment.getList());
-  }
-
-  @Override
-  public int hashCode() {
-    return getList().hashCode();
-  }
+  JsonObject toJson();
 
 }
