+++ src/main/java/io/vertx/ext/mail/impl/Utils.java
@@ -16,9 +16,15 @@
 
 package io.vertx.ext.mail.impl;
 
+import io.vertx.core.MultiMap;
+import io.vertx.core.http.CaseInsensitiveHeaders;
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
+
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -26,7 +32,7 @@ import java.util.Set;
 /**
  * @author <a href="http://oss.lehmann.cx/">Alexander Lehmann</a>
  */
-final class Utils {
+public final class Utils {
 
   /**
    * utility class only
@@ -80,6 +86,48 @@ final class Utils {
       // like [192.168.1.1] or [127.0.0.1]
       return "localhost";
     }
+  }
+
+  public static JsonObject multiMapToJson(final MultiMap headers) {
+    JsonObject json = new JsonObject();
+    for (String key : headers.names()) {
+      json.put(key, headers.getAll(key));
+    }
+    return json;
+  }
+
+  public static void putIfNotNull(final JsonObject json, final String key, final Object value) {
+    if (value != null) {
+      json.put(key, value);
+    }
+  }
+
+  public static MultiMap jsonToMultiMap(final JsonObject jsonHeaders) {
+    MultiMap headers = new CaseInsensitiveHeaders();
+    for (String key : jsonHeaders.getMap().keySet()) {
+      headers.add(key, getKeyAsStringOrList(jsonHeaders, key));
+    }
+    return headers;
+  }
+
+  @SuppressWarnings("unchecked")
+  public static List<String> getKeyAsStringOrList(JsonObject json, String key) {
+    Object value = json.getValue(key);
+    if (value == null) {
+      return null;
+    } else {
+      if (value instanceof String) {
+        return asList((String) value);
+      } else if (value instanceof JsonArray) {
+        return (List<String>) ((JsonArray) value).getList();
+      } else {
+        throw new IllegalArgumentException("invalid attachment type");
+      }
+    }
+  }
+
+  public static <T> List<T> asList(T element) {
+    return Collections.singletonList(element);
   }
 
 }
